Add AnimatedProgressBar with gradient
